Add step renaming and step groups
Added the ability to rename steps and an a way to have more than one scene. Scenes, or Groups, have their own set of steps and can be renamed.

Also changed various other parts related to saving/loading to accommodate step names and groups

(useAsync didn't work for some reason on my side, so I replaced that in FileDialog)